Move game logic outside of react component
